sunstudio compiler compatibility fixups
git-svn-id: https://segomo.elka.pw.edu.pl/svn/mrrocpp/base/trunk@2698 4c39dba3-d726-0410-8099-c98d187771e6

+++ src/ecp/common/ecp_g_smooth.cc
@@ -44,9 +44,9 @@ void smooth::set_absolute(void){
 
 bool smooth::eq(double a, double b)
 {
-	const double EPS = 0.0001;
+	const double EPS = 0.0001f;
 	const double diff = a - b;
-	return fabs(diff < EPS);
+	return fabs((double) (diff < EPS));
 }
 
 void smooth::generate_next_coords (void)
